Move all_options out to SearchEverything
